EDA of how big banks' stock progress during crisis
This project is an exploratory project about how the stock price of US Big 6 Banks progress during crisis until 2016.

This project and the resulting output are by no means a financial recommendation or advise, but rather just an exercise to improve my data exploratory skill,

The data used is a pickle data called all_banks (included in this repo), which contains Big6 open, close, high, low, volume stock detail. The aim of the project was to observe which bank suffered the most during crisis (indicated by lowest return value) and how the stock price performs after that. 

At the end of this project, I tried to generate candle stick and moving average using cufflinks library.